separated components & programs delete/archive modal
